Returner geografiskTilknytningDto
