+++ src/main/java/no/nav/familie/integrasjoner/geografisktilknytning/GeografiskTilknytningController.kt
@@ -4,11 +4,9 @@ import no.nav.familie.integrasjoner.client.rest.PdlRestClient
 import no.nav.familie.kontrakter.felles.PersonIdent
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.Ressurs.Companion.success
-import no.nav.familie.kontrakter.felles.arbeidsfordeling.Enhet
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import javax.validation.constraints.NotNull
 
 @RestController
 @RequestMapping("/api/arbeidsfordeling")
@@ -20,7 +18,7 @@ class GeografiskTilknytningController(
     @PostMapping("/geografisk-tilknytning/{tema}")
     fun hentGeografiskTilknytning(@PathVariable(name = "tema") tema: Tema,
                                              @RequestBody personIdent: PersonIdent)
-            : ResponseEntity<Ressurs<String>> {
+            : ResponseEntity<Ressurs<GeografiskTilknytningDto>> {
         return ResponseEntity.ok(success(pdlRestClient.hentGeografiskTilknytning(personIdent.ident, tema.toString())))
     }
