+++ src/main/java/no/nav/familie/integrasjoner/geografisktilknytning/GeografiskTilknytningDto.kt
@@ -1,0 +1,21 @@
+package no.nav.familie.integrasjoner.geografisktilknytning
+
+data class PdlHentGeografiskTilknytning(val hentGeografiskTilknytning: GeografiskTilknytningDto)
+
+data class GeografiskTilknytningDto(val gtType: GeografiskTilknytningType,
+                                    val gtKommune: String?,
+                                    val gtBydel: String?,
+                                    val gtLand: String?)
+
+
+enum class GeografiskTilknytningType {
+    KOMMUNE,
+    BYDEL,
+    UTLAND,
+    UDEFINERT
+}
+
+data class PdlGeografiskTilknytningRequest(val variables: PdlGeografiskTilknytningVariables,
+                                           val query: String)
+
+data class PdlGeografiskTilknytningVariables(val ident: String)
